+++ uppsrc/CodeEditor/LogSyntax.h
@@ -1,5 +1,5 @@
 class LogSyntax : public EditorSyntax {
 public:
 	virtual void            Highlight(const wchar *s, const wchar *end, HighlightOutput& hls,
-	                                  CodeEditor *editor, int line, int pos);
+	                                  CodeEditor *editor, int line, int64 pos);
 };
